+++ src/main/java/woowacourse/shoppingcart/domain/customer/Customer.java
@@ -3,6 +3,8 @@ package woowacourse.shoppingcart.domain.customer;
 import java.util.Objects;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import woowacourse.shoppingcart.domain.customer.address.FullAddress;
+import woowacourse.shoppingcart.domain.customer.password.Password;
+import woowacourse.shoppingcart.domain.customer.password.PasswordEncoderAdapter;
 import woowacourse.shoppingcart.domain.customer.privacy.Privacy;
 import woowacourse.shoppingcart.exception.DisagreeToTermsException;
 
@@ -35,7 +37,7 @@ public class Customer {
 
     public static Customer of(String email, String password, String profileImageUrl, Privacy privacy,
                               FullAddress fullAddress, boolean terms) {
-        return new Customer(new Email(email), Password.fromPlainText(password, new BCryptPasswordEncoder()),
+        return new Customer(new Email(email), Password.fromPlainText(password, new PasswordEncoderAdapter()),
                 new ProfileImageUrl(profileImageUrl),
                 privacy, fullAddress, terms);
     }
