+++ src/main/java/woowacourse/shoppingcart/domain/customer/password/Encoder.java
@@ -1,0 +1,7 @@
+package woowacourse.shoppingcart.domain.customer.password;
+
+public interface Encoder {
+    String encode(CharSequence rawPassword);
+
+    boolean matches(CharSequence rawPassword, String encodedPassword);
+}
